requirements.txt
flask
flask-mysql
flask-sqlalchemy
pymysql
config
python-dotenv